Fix fanout parameter mismatch
BitGoJS whitelisted `maxNumUnspentsToUse` for the fanout endpoint, but
Platform is expecting `maxNumInputsToUse` which also matches the
documentation.  The issue is fixed by updating the whitelist to use
`maxNumInputsToUse` instead.  Note that since the previously whitelisted
parameter in BitGoJS would never have resulted in the desired outcome,
this commit does not break working functionality of existing
integrations.

FIXES TICKET: BG-21100